## Changes
- Removed bottlenecks in ingestion
  - `create_datums`
  - `create_annotations`
  - `create_groundtruths`
  - `create_predictions`

## Results
On resource-constrained postgresql we see the following:

**Before**

```
"number_of_datums": 1000,
"ingest_runtime": "352.8 seconds",
"ingest_runtime": "350.6 seconds",
```

**After**

```
"number_of_datums": 1000,
"ingest_runtime": "15.8 seconds",
"ingest_runtime": "9.5 seconds",
```